% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioPathways}
\alias{bioPathways}
\title{Biological pathways collated from multiple databases.}
\format{
A data frame with 17,250 unique gene symbols and 14,598 unique biological pathways:
\describe{
  \item{Genesym}{Gene symbol}
  \item{Pathwayid}{Pathway Identifier}
  \item{Function}{Description of the pathway}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(bioPathways)
}
\description{
A data frame with information of the biological pathways that each gene symbol
is associated with, collated from multiple databases including Gene Ontology (GO)
Consortium, KEGG, Reactome Pathway database and ConsensusPathDB.
}
\keyword{datasets}
