% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{createInputParam}
\alias{createInputParam}
\title{Create Input parameter file for iOmicsPASS}
\usage{
createInputParam(
  data.X,
  data.Y = NULL,
  data.Z = NULL,
  within.net = NULL,
  btw.net = NULL,
  dir = "iOmicsPASS/inputFiles/",
  phenotype = NULL,
  pathway = NULL,
  standardize.data = T,
  log.transform = F,
  normalizeBy = NULL,
  min.obs = 5,
  min.prop = 0.8,
  knn.impute = F,
  knn.k = 10,
  max.block.impute = 1000,
  Cross.Validate = TRUE,
  num.Kfold = 5,
  min.thres = NULL,
  Enrichment = TRUE,
  tag = NULL,
  usePrior = FALSE,
  priorfile = NULL,
  bg.prop = 0.2,
  min.bg.size = 3,
  min.sig.size = 1
)
}
\arguments{
\item{data.X}{dataframe X}

\item{data.Y}{dataframe Y}

\item{data.Z}{dataframe Z}

\item{within.net}{network file for creating edges within X}

\item{btw.net}{network file for creating edges between X and Y}

\item{dir}{directory pointing to the location of the files}

\item{phenotype}{file with group sample information}

\item{pathway}{pathway file for network enrichment}

\item{standardize.data}{whether to standardize each data (default=TRUE)}

\item{log.transform}{whether to log-transform each data (default=FALSE)}

\item{normalizeBy}{Either "Y" or "Z". if "Y", coexpression scores in X are normalized by Y and if "Z", coexpression scores in X are normalized by Z.}

\item{min.obs}{minimum number of non-missing observations requires across each feature in each phenotypic outcome.}

\item{min.prop}{minimum proportion of non-missing observations requires across each feature in each phenotypic outcome.}

\item{knn.impute}{whether to perform K-nearest neighbor imputation for missing entries}

\item{knn.k}{number of folds for K-nearest neighbor imputation}

\item{max.block.impute}{number of blocks of samples to consider in KNN imputation}

\item{Cross.Validate}{Whether to run cross-validation (default=TRUE)}

\item{num.Kfold}{number of folds in CV}

\item{min.thres}{threshold to be used to derive the shrunken centroids in iOmicsPASS}

\item{Enrichment}{Whether to run network enrichment (default=TRUE)}

\item{tag}{a string that will be appended to the end of the output files like a tag.}

\item{usePrior}{whether to use an input prior to classify samples in discriminant model. If false, equal prior will be used.(default=FALSE)}

\item{priorfile}{file of the prior probabilities calculated from createPrior().}

\item{bg.prop}{minimum proportion of features in each pathway that are also present in the background list.}

\item{min.bg.size}{minimum number of features in each pathway that are also present in the background list.}

\item{min.sig.size}{minimum number of features in each pathway that are both part of the signature and present in the background list.}
}
\value{
a parameter file for input into iOmicsPASS.R()
}
\description{
Create Input parameter file for iOmicsPASS
}
\examples{
data(PhenotypeFile)
data(bioPathways)

## Running with estimated network from NetDeconvolute() ##
createInputParam(data.X="Combined_data.txt", within.net="Estimated_Network_glasso.txt",
phenotype =PhenotypeFile, Enrichment=FALSE)

## with Network-based pathway Enrichment ##
createInputParam(data.X="Combined_data.txt", within.net="Estimated_Network_glasso.txt",
phenotype =PhenotypeFile, pathway =bioPathways, Enrichment=TRUE)

data(Tulip_Protein)
data(Tulip_microRNA)
data(PPI_network)
data(TargetScan_network)

## Using known biological networks ##
createInputParam(data.X=Tulip_Protein,data.Y=Tulip_microRNA,phenotype =PhenotypeFile,
log.transform=TRUE,btw.net=TargetScan_network,within.net=PPI_network, Enrichment=FALSE)


}
