% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{createPrior}
\alias{createPrior}
\title{Create prior probabilities for discriminant model in iOmicsPASS}
\usage{
createPrior(
  FILE,
  y = NULL,
  var.cat = NULL,
  testFile = NULL,
  predict = FALSE,
  outputDir = "iOmicsPASS/inputFiles/",
  tag = NULL
)
}
\arguments{
\item{FILE}{dataframe with sample ID in the row names and clinical characteristics (for adjustment) in the columns.}

\item{y}{character string of the phenotypic or group variable.}

\item{var.cat}{a string or a vector of string matching the column names of FILE that are categorical variables.}

\item{testFile}{dataframe for test samples matching the colnames and variables in FILE}

\item{predict}{Whether or not to predict probabilities on a test dataset}

\item{outputDir}{directory to write the outputfile}

\item{tag}{a string that will be appended to the end of the output files like a tag.}
}
\value{
a dataframe with class probabilities assigned to each sample, used as prior in iOmicsPASS
}
\description{
Create prior probabilities for discriminant model in iOmicsPASS
}
\examples{
data(PhenotypeFile)
PhenotypeFile2 = PhenotypeFile
row.names(PhenotypeFile2) = PhenotypeFile$TulipID
PhenotypeFile2=PhenotypeFile2[,-1]
prior_out = createPrior(PhenotypeFile2, y = "Group",outputDir = "iOmicsPASS/inputFiles/")
prior_test = createPrior(PhenotypeFile2, y = "Group",predict=TRUE, testFile = TestData)
}
