% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{INSTALL.iOmicsPASS}
\alias{INSTALL.iOmicsPASS}
\title{Compile and build the C++ program for running iOmicsPASS}
\usage{
INSTALL.iOmicsPASS(currDir = NULL)
}
\arguments{
\item{currDir}{Current directory where iOmicsPASSv2plus is unzipped and where the makefile is.}
}
\description{
Compile and build the C++ program for running iOmicsPASS
}
\examples{
# set your working directory to iOmicsPASSv2plus folder #
INSTALL.iOmicsPASS()

}
