% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{iOmicsPASS.R}
\alias{iOmicsPASS.R}
\title{Carry out Predictive Analysis of Subnetwork Signatures}
\usage{
iOmicsPASS.R(
  ff = "input.param",
  outputDir = "iOmicsPASS/Output/",
  Cross.Validate = TRUE,
  plotCV = TRUE
)
}
\arguments{
\item{ff}{input parameter}

\item{outputDir}{directory to write the output files (default output to "iOmicsPASS/Output/")}

\item{Cross.Validate}{Whether or not cross-validation was performed in iOmicsPASS (default = TRUE)}

\item{plotCV}{Whether or not to plot the performance of the cross-validation (default = TRUE)}
}
\value{
dataframe with iOmicsPASS parameters used for prediction, multiple text files and plots written to output directory.
\itemize{
  \item AttributesTable.txt - Attributes table for every node in the network to be used in Cytoscape.
  \item BGlist.txt - A list of edges that are present in both the network and the input data.
  \item CVerrors.txt - A table of grid of threshold and their corresponding mean cross-validation errors and selected edges.
  \item CVplot_Penalty.pdf - A plot of the mean cross-validation error against the grid of threshold used to shrink the centroid.
  \item EdgesSelected_minThres.txt - dataframe with selected predictive edges and the dik scores for each phenotype that can be used to visualize directly in Cytoscape.
  \item Features_Neighbors.txt - Attributes table with neighboring node information for each node in the network.
  \item PredictiveEdges_Parameters.txt - A dataframe with with iOmicsPASS parameters used for prediction.
  \item SampleClass_Probabilities.txt - A dataframe with class probabilities assigned to the samples input.
  \item Ztransform_XXX.txt - Standardized data X/Y/Z.
  \item XXX_Enrichment_up.txt - a dataframe with the results of the enrichment of the selected edges (over-expressed) out of all the edges in the network for each class.
  \item XXX_Enrichment_down.txt - a dataframe with the results of the enrichment of the selected edges (under-expressed) out of all the edges in the network for each class.
}
}
\description{
Carry out Predictive Analysis of Subnetwork Signatures
}
\examples{
data(PhenotypeFile)
data(bioPathways)

## Running with estimated network from NetDeconvolute() ##
createInputParam(data.X="Combined_data.txt", within.net="Estimated_Network_glasso.txt",
phenotype =PhenotypeFile,Enrichment=FALSE)
iOmicsPASS.output<-iOmicsPASS.R(ff="input_param")

## pick optimal threshold as 2.4 and turn off CV ##
## rerun with network-based pathway enrichment ##
createInputParam(data.X="Combined_data.txt", within.net="Estimated_Network_glasso.txt",
phenotype=PhenotypeFile,pathway=bioPathways,Enrichment=TRUE,min.thres=2.4,Cross.Validate=FALSE)

iOmicsPASS.output<-iOmicsPASS.R(ff="input_param",Cross.Validate=FALSE, plotCV=FALSE)

data(Tulip_Protein)
data(Tulip_microRNA)
data(PPI_network)
data(TargetScan_network)

## Using known biological networks ##
createInputParam(data.X=Tulip_Protein,data.Y=Tulip_microRNA,phenotype=PhenotypeFile,
log.transform=TRUE,within.net=PPI_network,btw.net=TargetScan_network,Cross.Validate=TRUE,
Enrichment=FALSE, tag="KnownNetwork")
iOmicsPASS.output<-iOmicsPASS.R(ff="input_param_KnownNetwork")


}
