% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PhenotypeFile}
\alias{PhenotypeFile}
\title{Phenotype file for the Tulip study}
\format{
A two-column data frame with 17 subjects (row) and their phenotype:
\describe{
  \item{TulipID}{Subject identifier in the Tuplip study}
  \item{Group}{Phenotypic group of each subject. OIR for obese insulin resistant and LIS for lean insulin sensitive.}
  \item{Age}{Age of the subject}
  \item{BMI}{Body mass index of each subject}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(PhenotypeFile)
}
\description{
A data frame with two columns containing the phenotype assigned to the 18 subjects where
8 were obese insulin resistant (OIR, HOMA-IR >2.5) and 9 were lean insulin-sensitive
(LIS, HOMA-IR <1.0) normoglycemic males.
}
\keyword{datasets}
