% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Plot_CrossValidation}
\alias{Plot_CrossValidation}
\title{Plot Cross-validation performance}
\usage{
Plot_CrossValidation(outputdir)
}
\arguments{
\item{outputdir}{directory pointing to CVerrors.txt.}
}
\value{
a PDF file with the performance plot
}
\description{
Plot Cross-validation performance
}
\examples{
Plot_CrossValidation("iOmicsPASS/output/")
}
