% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPI_network}
\alias{PPI_network}
\title{Protein-Protein interaction network file}
\format{
A data frame with 1,201 unique proteins containing 1,499 interactions:
\describe{
  \item{geneA_genesym}{Gene symbol of Protein A}
  \item{geneB_genesym}{Gene symbol of Protein B that is interacting with Protein A}
  \item{sign}{Direction of protein-protein interaction, assumed to be all positive (indicated by 1)}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(PPI_network)
}
\description{
A data frame with three columns where the first two columns contain the pairs of protein identifiers
that chemically/physically interact with each other. The third column indicates the sign of the direction
of that interaction, which are assumed to be all positive interactions here.
}
\keyword{datasets}
