% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TargetScan_network}
\alias{TargetScan_network}
\title{microRNA to target gene network file}
\format{
A data frame with 882 unique genes and 125 unique microRNAs with 8,533 microRNA-gene targets:
\describe{
  \item{GeneSym}{Gene symbol}
  \item{miRNA}{microRNA probe ID}
  \item{sign}{Direction of miRNA-gene regulation where positive regulation of gene translation are indicated by 1 and inhibition of translation by -1.}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(TargetScan_network)
}
\description{
A dataframe with three columns where the first two columns represent the miRNA probes and their genes targets.
The third column indicates the direction of interaction, where microRNAs that are translation
inhibitor of its target genes as negative and those that regulates gene translation as positive.
}
\keyword{datasets}
