% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tulip_microRNA}
\alias{Tulip_microRNA}
\title{microRNA data in Tulip study.}
\format{
A data frame with 263 microRNA probes (rows) across 17 subjects (columns):
\describe{
  \item{miRNA}{microRNA probes}
  \item{TulipXX}{Sample ID in Tulip study}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(Tulip_microRNA)
}
\description{
A dataset containing 263 normalized microRNA copy number using multiplex RT-qPCR platform, MiRXES.
Original data contains 368 microRNA probes and only those that were different (p<0.1) between insulin
resistant and insulin sensitive individuals using 2-sample t-test were included in this example data.
}
\keyword{datasets}
