% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tulip_Protein}
\alias{Tulip_Protein}
\title{Protein expression data in Tulip study. Original data contains 1,499 proteins quantified and only
those that were different (p<0.1) between insulin resistant and insulin sensitive individuals
using 2-sample t-test were included in this example data.}
\format{
A data frame with 1,499 proteins (rows) across 17 subjects (columns):
\describe{
  \item{Protein}{Protein symbol}
  \item{TulipXX}{Sample ID in Tulip study}
}
For further details, see \url{https://pubmed.ncbi.nlm.nih.gov/31024340/}
}
\usage{
data(Tulip_Protein)
}
\description{
A dataset containing the 266 plasma protein abundance quantified by LC-MS across
17 males.
}
\keyword{datasets}
